moved Settings struct to the library crate
