+++ src/cli_args/mod.rs
@@ -1,39 +1,8 @@
 use clap::builder::styling::{AnsiColor, Color, Style, Styles};
 use clap::{arg, command, ColorChoice, Parser};
-use config::{Config, ConfigError, File, ValueKind};
-use idf_im_lib::get_log_directory;
-use log::{debug, info, LevelFilter};
-use serde::{Deserialize, Serialize};
-use std::fs::OpenOptions;
-use std::io::Write;
 use std::path::PathBuf;
-use toml::value::Array;
-
-use log4rs::{
-    append::{console::ConsoleAppender, file::FileAppender},
-    config::{Appender, Root},
-    encode::pattern::PatternEncoder,
-};
 
 const VERSION: &str = env!("CARGO_PKG_VERSION");
-
-#[derive(Debug, Deserialize, Default, Serialize)]
-pub struct Settings {
-    pub path: Option<PathBuf>,
-    pub idf_path: Option<PathBuf>,
-    pub tool_download_folder_name: Option<String>,
-    pub tool_install_folder_name: Option<String>,
-    pub target: Option<Vec<String>>,
-    pub idf_versions: Option<Vec<String>>,
-    pub tools_json_file: Option<String>,
-    pub idf_tools_path: Option<String>,
-    pub config_file: Option<PathBuf>,
-    pub non_interactive: Option<bool>,
-    pub wizard_all_questions: Option<bool>,
-    pub mirror: Option<String>,
-    pub idf_mirror: Option<String>,
-    pub recurse_submodules: Option<bool>,
-}
 
 fn custom_styles() -> Styles {
     Styles::styled()
@@ -69,7 +38,7 @@ pub struct Cli {
     )]
     path: Option<String>,
     #[arg(short, long, value_name = "FILE")]
-    config: Option<PathBuf>,
+    pub config: Option<PathBuf>,
 
     #[arg(
         short,
@@ -86,38 +55,38 @@ pub struct Cli {
     idf_versions: Option<String>,
 
     #[arg(long)]
-    tool_download_folder_name: Option<String>,
+    pub tool_download_folder_name: Option<String>,
 
     #[arg(long)]
-    tool_install_folder_name: Option<String>,
+    pub tool_install_folder_name: Option<String>,
 
     #[arg(
         long,
         help = "Path to tools.json file relative from ESP-IDF installation folder"
     )]
-    idf_tools_path: Option<String>,
+    pub idf_tools_path: Option<String>,
 
     #[arg(
         long,
         help = "Path to idf_tools.py file relative from ESP-IDF installation folder"
     )]
-    tools_json_file: Option<String>,
+    pub tools_json_file: Option<String>,
 
     #[arg(short, long)]
-    non_interactive: Option<bool>,
+    pub non_interactive: Option<bool>,
 
     #[arg(
         short,
         long,
         help = "url for download mirror to use instead of github.com"
     )]
-    mirror: Option<String>,
+    pub mirror: Option<String>,
 
     #[arg(
         long,
         help = "url for download mirror to use instead of github.com for downloading esp-idf"
     )]
-    idf_mirror: Option<String>,
+    pub idf_mirror: Option<String>,
 
     #[arg(
         short,
@@ -125,132 +94,20 @@ pub struct Cli {
         action = clap::ArgAction::Count,
         help = "Increase verbosity level (can be used multiple times)"
     )]
-    verbose: u8,
+    pub verbose: u8,
 
     #[arg(short, long, help = "Set the language for the wizard (en, cn)")]
-    locale: Option<String>,
+    pub locale: Option<String>,
 
     #[arg(long, help = "file in which logs will be stored (default: eim.log)")]
-    log_file: Option<String>,
+    pub log_file: Option<String>,
 
     #[arg(
         short,
         long,
-        help = "Should the installer recurse into submodules of the ESP-IDF repository (derfault true) "
+        help = "Should the installer recurse into submodules of the ESP-IDF repository (default true) "
     )]
-    recurse_submodules: Option<bool>,
-}
-
-impl Settings {
-    pub fn new() -> Result<Self, ConfigError> {
-        let cli = Cli::parse();
-
-        Self::setup_logging(&cli)?;
-        Self::set_locale(&cli.locale);
-
-        let mut builder = Config::builder()
-            .add_source(File::with_name("config/default").required(false))
-            .add_source(File::with_name("config/development").required(false));
-
-        if let Some(config_path) = cli.config.clone() {
-            debug!("Using config file: {}", config_path.display());
-            builder = builder.add_source(File::from(config_path));
-        }
-
-        builder = builder.add_source(config::Environment::with_prefix("ESP").separator("_"));
-
-        for (key, value) in cli.into_iter() {
-            if let Some(v) = value {
-                if key != "config" {
-                    debug!("Setting {} to {:?}", key, v);
-                    builder = builder.set_override(key, v)?;
-                }
-            }
-        }
-        let cfg = builder.build()?;
-        cfg.try_deserialize()
-    }
-
-    fn setup_logging(cli: &Cli) -> Result<(), ConfigError> {
-        let log_file_name = cli.log_file.clone().map_or_else(
-            || {
-                get_log_directory()
-                    .map(|dir| dir.join("eim.log"))
-                    .unwrap_or_else(|| {
-                        eprintln!("Failed to get log directory, using default eim.log");
-                        PathBuf::from("eim.log")
-                    })
-            },
-            PathBuf::from,
-        );
-
-        let logfile = FileAppender::builder()
-            .encoder(Box::new(PatternEncoder::new("{d} - {l} - {m}\n")))
-            .build(log_file_name)
-            .map_err(|e| ConfigError::Message(format!("Failed to build file appender: {}", e)))?;
-
-        let stdout = ConsoleAppender::builder()
-            .encoder(Box::new(PatternEncoder::new("{d} - {l} - {m}\n")))
-            .build();
-
-        let log_level = match cli.verbose {
-            0 => LevelFilter::Info,
-            1 => LevelFilter::Debug,
-            _ => LevelFilter::Trace,
-        };
-
-        let config = log4rs::Config::builder()
-            .appender(
-                Appender::builder()
-                    .filter(Box::new(log4rs::filter::threshold::ThresholdFilter::new(
-                        LevelFilter::Trace,
-                    )))
-                    .build("file", Box::new(logfile)),
-            )
-            .appender(
-                Appender::builder()
-                    .filter(Box::new(log4rs::filter::threshold::ThresholdFilter::new(
-                        log_level,
-                    )))
-                    .build("stdout", Box::new(stdout)),
-            )
-            .build(
-                Root::builder()
-                    .appender("stdout")
-                    .appender("file")
-                    .build(LevelFilter::Trace),
-            )
-            .map_err(|e| ConfigError::Message(format!("Failed to build log4rs config: {}", e)))?;
-
-        log4rs::init_config(config)
-            .map_err(|e| ConfigError::Message(format!("Failed to initialize logger: {}", e)))?;
-
-        Ok(())
-    }
-
-    fn set_locale(locale: &Option<String>) {
-        match locale {
-            Some(l) => {
-                rust_i18n::set_locale(l);
-                info!("Set locale to: {}", l);
-            }
-            None => debug!("No locale specified, defaulting to en"),
-        }
-    }
-
-    pub fn save(&self, file_path: &str) -> Result<(), ConfigError> {
-        let toml_value = toml::to_string(self).map_err(|e| ConfigError::Message(e.to_string()))?;
-        let mut file = OpenOptions::new()
-            .write(true)
-            .truncate(true)
-            .create(true)
-            .open(file_path)
-            .map_err(|e| ConfigError::Message(e.to_string()))?;
-        file.write_all(toml_value.as_bytes())
-            .map_err(|e| ConfigError::Message(e.to_string()))?;
-
-        Ok(())
-    }
+    pub recurse_submodules: Option<bool>,
 }
 
 impl IntoIterator for Cli {
